Source/WebCore: Patch by Abhishek Arya <inferno@chromium.org> on 2011-07-13 Reviewed by Adam Barth.
Issue with Frame lifetime due to deletion in beforeload event.
https://bugs.webkit.org/show_bug.cgi?id=64457

Copy the Frame protector higher in the stack from loadWithDocumentLoader
to loadFrameRequest since any of loadPostRequest or loadURL can call
loadWithDocumentLoader, thereby dispatching the beforeload event and
blowing away the frame. This deleted frame will be later accessed in
the loadFrameRequest function causing a crash.

Test: fast/events/form-iframe-target-before-load-crash2.html

* loader/FrameLoader.cpp:
(WebCore::FrameLoader::loadFrameRequest):
(WebCore::FrameLoader::loadWithDocumentLoader):

LayoutTests: Tests that we do not crash when frame is blown away in a beforeload
event.
https://bugs.webkit.org/show_bug.cgi?id=64457

Reviewed by Adam Barth.

* fast/events/form-iframe-target-before-load-crash.html:
* fast/events/form-iframe-target-before-load-crash2-expected.txt: Added.
* fast/events/form-iframe-target-before-load-crash2.html: Added.


Identifier: 80193@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@90936 268f45cc-cd09-0410-ab3c-d52691b4dbfc