Specify a cascade list consisting of the last resort font for Core Text https://bugs.webkit.org/show_bug.cgi?id=64747
Reviewed by Sam Weinig.

No new tests, because this does not affect behavior.

This prevents Core Text from taking its default, longer fallback list when the primary font does
not include a character. This is OK to do because WebKit never uses the results of Core Text
fallback anyway.

* platform/graphics/cocoa/FontPlatformDataCocoa.mm:
(WebCore::cascadeToLastResortFontDescriptor): Added. Returns a CTFontDescriptor with a cascade
list consisting of the last resort font.
(WebCore::FontPlatformData::ctFont): Changed to include the cascadeToLastResortFontDescriptor
in the returned font.



Identifier: 80457@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91209 268f45cc-cd09-0410-ab3c-d52691b4dbfc