https://bugs.webkit.org/show_bug.cgi?id=64809 REGRESSION (r91114-r91116): fast/dom/error-to-string-stack-overflow.html failing on Windows 7 Release (Tests)
Reviewed by Adam Roben.

I think this just needs new results. The test deliberately triggers a stack overflow,
and doesn't catch it, so the error is printed to the console. Previously RangeError
had the wrong prototype, and didn't convert the error into a useful error message,
printing a blank line in the console. Since r91116 we inherit from Error correctly,
and print the error to the console.

* fast/dom/error-to-string-stack-overflow-expected.txt:
* fast/js/script-tests/object-prototype-properties.js:



Identifier: 80548@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91305 268f45cc-cd09-0410-ab3c-d52691b4dbfc