show a list of average test times in the treemap https://bugs.webkit.org/show_bug.cgi?id=64899
Reviewed by Adam Roben.

For now, you click a link and it replaces the treemap
with a reverse sorted list of directories based on
average test runtime within that directory.

To cut some of the noise, we don't show directories
that only have one test in them or where the average
runtime is less than 100ms.

* TestResultServer/static-dashboards/treemap.html:

Identifier: 80644@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91415 268f45cc-cd09-0410-ab3c-d52691b4dbfc