Cannot click to position the caret to the right of an image with display:block style. https://bugs.webkit.org/show_bug.cgi?id=64962 <rdar://problem/9446343>
Reviewed by David Hyatt.

Source/WebCore:

Test: editing/selection/click-on-block-image.html

The current implementation of positionForPoint always returns the visible
position to the left of the image, if the image has a block style.
This patch changes the behavior for block styled images to match the behavior
for inline images.

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::positionForPoint): Modified to avoid relying entirely
on InlineBox to calculate the position.

LayoutTests:

* editing/selection/click-on-block-image-expected.txt: Added.
* editing/selection/click-on-block-image.html: Added.



Identifier: 80700@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91481 268f45cc-cd09-0410-ab3c-d52691b4dbfc